Add files via upload
This Tableau Project has charts of Total Sales, Profits, and Quantity, along with comparisons of different years of the particular company. I have used some specific colours, fonts and designs to make it easier for the users to access the Project and understand the company's position. In this project, the user can find Bans and Charts and can also access their preferred year of the company by using Filter on the Dashboard of the project. The CSV file which has been used in this project has also been uploaded to this repository.